Fix frame pacing and stabilize simulation time display.
Cap runtime to 60 FPS, enable vsync presentation, and format sim time in the HUD as mm:ss.t with higher-precision accumulation for steadier output.

Made-with: Cursor

+++ src/Game.h
@@ -65,7 +65,7 @@ private:
 
     D3DXVECTOR3 windmill_pos_{72.f, 0.f, -58.f};
 
-    float sim_time_ = 0.f;
+    double sim_time_ = 0.0;
     /** Exponential moving average of frame rate for HUD (updated in Update). */
     float fps_smoothed_ = 0.f;
     D3DCOLOR clear_color_ = D3DCOLOR_XRGB(135, 206, 235);
